backend/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-cors==4.0.0
flask-jwt-extended==4.6.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
